metrics: Split the pod deletion logic into a separate file
This is a non-functional change, only reorganising code for easier development.
